+++ CHANGELOG.md
@@ -4,8 +4,11 @@ OJP-Demo URL: https://opentdatach.github.io/ojp-demo-app/
 
 ----
 
+09.May 2025
+- Display Occupancy Level - [Occupancy in OJP 2.0 #223](https://github.com/openTdataCH/ojp-demo-app-src/issues/223), [PR #253](https://github.com/openTdataCH/ojp-demo-app-src/pull/225)
+
 02.May 2025
-- Catch `changeWithinVehicle` for Flügelzüge - [[PR #222](https://github.com/openTdataCH/ojp-demo-app-src/pull/222)
+- Catch `changeWithinVehicle` for Flügelzüge - [PR #222](https://github.com/openTdataCH/ojp-demo-app-src/pull/222)
   - catch TransferLeg `changeWithinVehicle` flag for OJP 2.0
   - refactor trip info to show information when the legs were merged